requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.2
django-ckeditor==6.5.1
django-formtools==2.4
django-js-asset==2.0.0
Pillow==9.5.0
pycodestyle==2.10.0
reportlab==3.6.12
sqlparse==0.4.3
tzdata==2023.3
